client::component: renamed loader.page => Loader
